requirements.txt
flask
numpy
scikit-learn
sklearn_pandas
imbalanced-learn
nltk
nltk.corpus
nltk.stem